Complete conversion of scrapers to plugins
