+++ bot.py
@@ -32,7 +32,7 @@ class Bot(discord.Client):
     WATCH_MSG_LOCATION='alwayswatchmsgloc'
     ROLE_MSG_LOCATION='rolemessagesloc'
 
-    def __init__(self, config, log, checks, stop_queue, always_watch_messages):
+    def __init__(self, config, log, stop_queue, always_watch_messages):
         '''(str, Logger, fun) -> Bot
         config: a string which is the loaction of the base config file
         log: a Logger for dumping info
@@ -43,8 +43,6 @@ class Bot(discord.Client):
             pass
         self.data_config = dataloader.datafile(config).content[DEFAULT]
         self.log = log
-        # TODO(14flash): Plugin refactor, where we won't need a doCheck() func anymore
-        self.checks = checks
         self.data = dict()
         self.commands = OrderedDict() # maps names to commands
         self.reactions = OrderedDict() # maps names to reaction commands
@@ -95,7 +93,6 @@ class Bot(discord.Client):
         self.reactions[name]=cmd
     @asyncio.coroutine
     def on_message(self, message):
-        yield from self.checks(self)
         yield from self.message_stuff()
         for cmd in self.commands:
             try:
@@ -134,28 +131,13 @@ class Bot(discord.Client):
 
     @asyncio.coroutine
     def on_ready(self):
+        print("Bot online & running startup (this may take a while)")
         self.log.info('API connection created successfully')
         self.log.info('Username: ' + str(self.user.name))
         #self.log.info('Email: ' + str(self.email))
         self.log.info('Connected to %s servers' %len(self.servers))
-        self.setup_channels()
         yield from self.load_messages()
-        yield from self.checks(self)
-
-    def setup_channels(self):
-        '''() -> None
-        Convinience fuction for on_ready()'''
-        for i in self.get_all_channels(): # play the matchy-matchy game with server names
-            if i.name == self.get_data(CHANNEL_LOC, 'twitter'):
-                self.twitterchannel = i
-                self.log.info('twitter channel found')
-            if i.name == self.get_data(CHANNEL_LOC, 'forum'):
-                self.forumchannel = i
-                self.log.info('forum channel found')
-            if i.name == self.get_data(CHANNEL_LOC, 'reddit'):
-                self.redditchannel = i
-                self.log.info('reddit channel found')
-        print("Channels found")
+        print("All messages loaded. Full functionality enabled")
 
     def load_messages(self):
         '''() -> None
