--- data/freeforums.txt
@@ -1,1 +1,0 @@
-most recent thread:http://ideahavers.freeforums.net/thread/93/why-knowledge-move-slowly-645
