--- data/reddit.config
@@ -1,4 +1,0 @@
-[DEFAULT]
-urlfilepath = ./data/redditwatchlist.txt
-datafilepath = ./data/reddit.txt
-period = 30
