--- data/reddit.txt
@@ -1,2 +1,0 @@
-/r/pcmasterrace/comments/7nthay/a_quick_summary_of_the_current_intel_cpu_bug/ds4hfzh/
-/r/pcmasterrace/comments/7nthay/a_quick_summary_of_the_current_intel_cpu_bug/ds53b8y/
