--- data/twitter.config
@@ -1,4 +1,0 @@
-[DEFAULT]
-url = https://twitrss.me/twitter_user_to_rss/?user=openideaproject
-datafilepath = ./data/twitter.txt
-period = 3000
