--- data/twitter.txt
@@ -1,2 +1,0 @@
-most recent tweet:https://twitter.com/openideaproject/status/1006294941943943168
-second most recent tweet:https://twitter.com/SUEtheTrex/status/1004814892854140929
