+++ main.py
@@ -28,11 +28,6 @@ DEV_PERM = 'devperm'
 STATISTICS_PERM = 'statsperm'
 MANAGE_ROLES_PERM = 'manageroleperm'
 
-TWITTER_URL = r"https://twitter.com/openideaproject"
-TWITTER_PROFILE_ICON = r"https://pbs.twimg.com/profile_images/913153081566691328/wDhS7B1w_400x400.jpg"
-TWITTER_LOGO = r"https://pbs.twimg.com/profile_images/875087697177567232/Qfy0kRIP_400x400.jpg"
-REDDIT_LOGO = r"https://pbs.twimg.com/profile_images/868147475852312577/fjCSPU-a_400x400.jpg" #ignore that this is a Twitter link *whistles innocently*
-
 def configureDiscordLogging():
     '''() -> None
     set ups discord log so that it outputs to ./discord.log'''
@@ -109,43 +104,6 @@ class CustomNamespace:
     '''For creating custom namespaces wherever necessary'''
     pass
 
-@asyncio.coroutine
-def doChecks(bot):
-    '''(discord.Client) -> None
-    checks to make sure no messages need to be sent about something special, like scraper updates'''
-    pass
-    '''
-    threads = list()
-    while not qForum.empty():
-        threads.append(qForum.get())
-    for thread in threads[::-1]:
-        #thread = qForum.get()
-        #yield from bot.send_message(bot.forumchannel, "BEEP BOOP. This is a debug msg:"+str(thread[1]))
-        users = []
-        for i in thread[1]:
-            user = matchuser(i)
-            if user != None and user not in users:
-                users.append(user)
-        if len(users)>0:
-            yield from bot.send_message(bot.forumchannel, mentionChain(users), embed=embed.create_embed(description="In: "+thread[0], author={"name" : thread[1][-1]["name"], "url" : FORUM_URL+thread[1][-1]["url"], "icon_url" : None}, footer={"text":"Forum", "icon_url":None}))
-        else:
-            yield from bot.send_message(bot.forumchannel, embed=embed.create_embed(description="In: "+thread[0], author={"name" : thread[1][-1]["name"], "url" : FORUM_URL+thread[1][-1]["url"], "icon_url" : None}, footer={"text":"Forum", "icon_url":None}))
-    tweets = list()
-    while not qTwitter.empty():
-        tweets.append(qTwitter.get())
-    for tweet in tweets[::-1]:
-        #tweet = qTwitter.get()
-        if not tweet["retweet"]:
-            yield from bot.send_message(bot.twitterchannel, embed=embed.create_embed(author={"name":tweet["author"], "url":tweet["url"], "icon_url":TWITTER_PROFILE_ICON}, description=tweet["content"], footer={"text":"Twitter", "icon_url":TWITTER_LOGO}))
-        else:
-            yield from bot.send_message(bot.twitterchannel, embed=embed.create_embed(author={"name":"openideaproject retweeted "+tweet["author"], "url":tweet["url"], "icon_url":TWITTER_PROFILE_ICON}, description=tweet["content"], footer={"text":"Twitter", "icon_url":TWITTER_LOGO}))
-    comments = list()
-    while not qReddit.empty():
-        comments.append(qReddit.get())
-    for comment in comments[::-1]:
-        #comment = qReddit.get()
-        yield from bot.send_message(bot.redditchannel, embed=embed.create_embed(description="A comment has been posted here: " + comment[0] + " (direct link: "+comment[1]+" )", footer={"text":"Reddit", "icon_url":REDDIT_LOGO}))'''
-
 if __name__ == '__main__':
     # main
     # init stuff
@@ -154,10 +112,6 @@ if __name__ == '__main__':
     config.content = config.content["DEFAULT"]
     credentials = dataloader.datafile(config.content["credentialsloc"])
     credentials.content = credentials.content["DEFAULT"]
-    channels = dataloader.datafile(config.content["channelsloc"])
-    channels.content = channels.content["DEFAULT"]
-    forumdiscorduser = dataloader.datafile(config.content["forumdiscorduserloc"])
-    forumdiscorduser.content = forumdiscorduser.content["DEFAULT"]
 
     log = mainLogging()
 
@@ -168,8 +122,8 @@ if __name__ == '__main__':
     commands = dict()
     reactions = dict()
 
-    print("initializing bot...")
-    bot = botlib.Bot("./data/config.config", log, doChecks, stop, always_watch_messages)
+    print("Initializing bot...")
+    bot = botlib.Bot("./data/config.config", log, stop, always_watch_messages)
     bot.add_data(botlib.CHANNEL_LOC)
 
     # variables that should be accessible to all other commands are declared here
@@ -181,10 +135,6 @@ if __name__ == '__main__':
     role_messages = savetome.load_role_messages(config.content[ROLE_MSG_LOCATION], all_emojis_func)
     bot.role_messages = role_messages
 
-    '''
-    qForum = Queue()
-    qTwitter = Queue()
-    qReddit = Queue()'''
     global qRedditURLAdder
     qRedditURLAdder = Queue()
 
@@ -211,14 +161,6 @@ if __name__ == '__main__':
     # register plugins
     for plugin_name in plugins:
         bot.register_plugin(plugins[plugin_name], plugin_name)
-
-    '''
-    forumScraper = Process(target = scraperff.continuousScrape, args = (qForum, stop, ))
-    forumScraper.start()
-    twitterScraper = Process(target = scrapert.continuousScrape, args = (qTwitter, stop, ))
-    twitterScraper.start()
-    redditScraper = Process(target = scraperred.continuousScrape, args = (qReddit, stop, qRedditURLAdder, ))
-    redditScraper.start()'''
 
     # bot.register_command(invalid.InvalidCommand(user=user_func, invalid_message=config.content["invalidmessagemessage"]))
     if "token" in credentials.content:
@@ -252,9 +194,5 @@ if __name__ == '__main__':
         bot.plugins[cmd_name]._shutdown()
 
     savetome.save_role_messages(config.content[ROLE_MSG_LOCATION], role_messages)
-    '''
-    twitterScraper.join()
-    forumScraper.join()
-    redditScraper.join()'''
 
     print("Ended")
