+++ plugins/scrapers/scraperred.config
@@ -1,6 +1,6 @@
 [DEFAULT]
-url = http://ideahavers.freeforums.net/rss/public
-datafilepath = ./plugins/scrapers/freeforums.txt
+urlfilepath = ./plugins/scrapers/redditwatchlist.txt
+datafilepath = ./plugins/scrapers/reddit.txt
 period = 30
 threadedperiod = 300
 # valid options are join, terminate, none and custom [and kill, Python 3.7+ only]
