+++ plugins/scrapers/scrapert.config
@@ -1,6 +1,6 @@
 [DEFAULT]
-url = http://ideahavers.freeforums.net/rss/public
-datafilepath = ./plugins/scrapers/freeforums.txt
+url = https://twitrss.me/twitter_user_to_rss/?user=openideaproject
+datafilepath = ./plugins/scrapers/twitter.txt
 period = 30
 threadedperiod = 300
 # valid options are join, terminate, none and custom [and kill, Python 3.7+ only]
